projects/01-fastapi-minimal-chat/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.99.9
python-dotenv==1.0.1
